Added message APIs

+++ backend/apis/group-clear.js
@@ -19,12 +19,21 @@ module.exports = {
 // Contol //
 ////////////
 
+/*
+Clears out all messages, channels, and waypoints for the given group
+request			must contain groupToClear, which must match the json web token
+*/
 function groupClearPreAuth(request, response) {
 	log.line("Clear group", 1);
 	gatekeeper.gate(request, response, groupClear);
 }
 
 function groupClear(request, response, token) {
+	if (process.env.NODE_ENV !== "test" && process.env.NODE_ENV !== "test-log") {
+		log.__("Cannot clear a group when not in test mode");
+		response.sendStatus(403);
+		return;
+	}
 	if (request.body.groupToClear && request.body.groupToClear === token.group) {
 		deleteAllChannels(request, response);
 	} else if (!request.body.groupToClear) {
@@ -43,23 +52,64 @@ function deleteAllChannels(request, response) {
 	const query = data.store().createQuery("channel")
 	.hasAncestor(data.store().key(["group", request.body.groupToClear]))
 	.select("__key__");
-	data.store().runQuery(query, (error, results) => { deleteChannels(error, results, response); });
+	data.store().runQuery(query, (error, results) => { deleteChannels(error, results, request, response); });
 }
 
-function deleteChannels(error, channelsToDelete, response) {
+function deleteChannels(error, channelsToDelete, request, response) {
 	const keys = channelsToDelete.map((channel) => {
 		return channel[data.store().KEY];
+	});
+	data.store().delete(keys, (error) => { deleteAllMessages(error, request, response); });
+}
+
+function deleteAllMessages(error, request, response) {
+	if (error) {
+		log.error("Could not delete channels");
+		log.error(error);
+		response.sendStatus(500);
+		return;
+	}
+	const query = data.store().createQuery("message")
+	.filter("group", "=", request.body.groupToClear)
+	.select("__key__");
+	data.store().runQuery(query, (error, results) => { deleteMessages(error, results, request, response); });
+}
+
+function deleteMessages(error, messagesToDelete, request, response) {
+	const keys = messagesToDelete.map((message) => {
+		return message[data.store().KEY];
+	});
+	data.store().delete(keys, (error) => { deleteAllWaypoints(error, request, response); });
+}
+
+function deleteAllWaypoints(error, request, response) {
+	if (error) {
+		log.error("Could not delete messages");
+		log.error(error);
+		response.sendStatus(500);
+		return;
+	}
+	const query = data.store().createQuery("waypoint")
+	.filter("group", "=", request.body.groupToClear)
+	.select("__key__");
+	data.store().runQuery(query, (error, results) => { deleteWaypoints(error, results, request, response); });
+}
+
+function deleteWaypoints(error, waypointsToDelete, request, response) {
+	const keys = waypointsToDelete.map((waypoint) => {
+		return waypoint[data.store().KEY];
 	});
 	data.store().delete(keys, (error) => { respond(error, response); });
 }
 
 function respond(error, response) {
 	if (error) {
-		log.line("Could not delete channels", "error");
+		log.line("Could not delete waypoints", "error");
 		log.line(error, "error");
 		response.sendStatus(500);
 		return;
 	}
+	log.line("Success", 2);
 	response.status(200).send(JSON.stringify({
 		success: true
 	}));
